tweak code based on @toomuchdesign feedbacks
